Add calc command
This command will calculate an infix expression by converting it to
RPN using the shunting yard algorithm and then simply interpreting
that. The implementation is a bit divergent from the default
shuntingyard algorithm, particularly in the implementation of the
parentheses and function. These are implemented by calling interpret
recursively instead of pushing the parentheses on the operator stack
during execution.

The function currently supported are sin/1, cos/1, tan/1, asin/1,
acos/1, atan/1, ln/1, sqrt/1, log/2, max/2. The operators supported
are ^, *, /, +, -.